changing HomeComponent path after moving it in Components component
